Support for external process supervisor
This commit adds support for interfacing with an external process
supervisor.  This is common in more safety critical applications
where liveness of processes' main loop must be verified.

No extra command line option is needed to activate support for an
external kick.  Simply use SIGUSR1 to assume control of the wdt and
continuously kick it from another process.  Use SIGUSR2 to signal
graceful exit, in which case watchdogd will resume using its built-in
kick.

If SIGUSR2 is not used when the supervisor exits, or if SIGUSR1 is not
continuously received, the system will reboot as expected when the HW
wdt timeout is reached.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>