Support for forcing external kick, with optional startup delay
This commit adds support for forcing an external kick, with an optional
startup delay.  Very useful in setups where an external supervisor
daemon is expected to use SIGUSR1 to watchdogd to kick the hardware
watchdog.

In this mode we expect there is an external daemon and if that daemon
does not start and not even sends an initial SIGUSR1 we can catch that
too.  The option supports a startup delay feature to work around startup
phenomena to will and do occur.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>