Add pmon to default build, enable with command line switch.
This patch no only adds pmon to the default build, it also makes it
possible to adjust the real-time priority pmon, and watchdogd, runs
at when supervising processes.

When pmon is enabled but not yet actively monitoring any subscribed
process watchdogd runs as SCHED_OTHER prio 0.  But as soon as one
process subsbscribes watchdogd changes to SCHED_RR prio 98, by default.
This RT priority value can be changed with the command line argument

    watchdogd -p PRIO
    watchdogd --pmon=PRIO

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>