Remove support for external supervisor (SIGURS1 & SIGUSR2)
With the addition of the PMON API watchdogd is a standalone system, and
process supervisor.  Hence, the old SIGUSR1 + SIGUSR2 "API" for hooking
in an external supervisor is no longer needed.

Also, with the freeing up of `-x` the `--safe-exit` short option `-e` is
replaced as well in this commit.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>