Encapsulate test mode and untangle it from libwdog
Remove `__wdog_testmode` and separate api_init() from libwdog.  The API
initialization is now handled separately for server and client to reduce
complexity of the code.

In the client code we can check for the existance of the socket file to
determine if we are in regular or test mode.  If neither, then watchdogd
has not yet started.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>