watchdogd: Probe WDT chipset model and capabilities
To make watchdogd more useful to a wider audience we need to support
more limited chipsets.  This patch extends the internal API of the
wdt_get_bootstatus() function to probe for the WDT chipset model, its
capabilityies and also adds a wdt_capability() predicate function to
be used in later commits.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>