watchdogd: If WDT doesn't support WDIOF_CARDRESET, log kick to syslog
A WDT that doesn't support WDIOF_CARDRESET cannot tell us if we rebooted
due to us having failed to kick it in time.  At offline analysis it is
helpful if multiple sources of information can be consulted to determine
the actual reset cause.

If the log message this patch adds is annoying, the user can simply
change the log level of watchdogd.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>