watchdogd: Improve WDT status titles, had confusingly similar names
This patch hopefully makes it more clear to read.  E.g., the WDT boot
status (flags) are not really the reset cause.  Instead call it what it
really is, the boot status.  Also, the "Reason" is the human readable
string version of the Reset cause, so let's call it that.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>