Add wdog_reboot_reason_raw() API, for compat
The wdog_reboot_reason_raw() reads the actual backend store for the
currently stored reboot reason, which could be the next reboot reason
in case of sudden reboots or power-loss.

The wdog_reboot_reason() returns the reboot reason read at boot, or when
watchdogd first started.  If watchdogd should be restarted these API's
may return the same data.  However, in production we do not want to
allow the daemon to be restarted, the WDT should have nowayout set.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>