Add fallback handling to compat API wdog_forced_reset()
This patch adds fallback reset handling to the wdog_forced_reset()
compat API.  In case communication with watchdogd fails we attempt
to fall back to telling init to reboot, and if that fails we take
the last resort and force a kernel reboot.

Also, the original API does not return int, but is void.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>