Prepare for shutdown/reboot on SIGPWR, force reboot on SIGTERM
If a user (init) sends SIGPWR we take this as heads-up to a shutdown or
reboot, of this is followed by SIGTERM we reboot immediately using WDT.

The user may also send SIGPWR twice, instead of a SIGTERM, the effect
will be the same.

After SIGPWR has been received watchdogd has stored the reboot reason
and started a 10 sec countdown timer which will force reboot should
the system hang before sending the final SIGTERM (or second SIGPWR).

This change is for the cooperative reboot handling with Finit.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>