Replace "PMON" with "process supervisor" in all text and code
The name PMON was introduced by myself as a sort of "concept" ... but
since there is no common knowledge in the rest of the world what a
"PMON" is, I've decided it's better to just call it by its proper name.

In text "PMON" is replaced with process supervisor, I may mention use
the work "monitor" as well, but only in passing.

In code "PMON" and "pmon" is either removed entirely or replaced with
just "SUPERVISOR", or "supervisor".  This because it is often used in
conjunction with a "WDOG" or "WATCHDOG" prefix.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>