Perform forced/controlled reset using Finit instead
If we run on a system with Finit we can ask it to do an orderly reboot,
stop services (allow dying gasp from e.g. snmpd), sync and do a proper
unmount of all file systems.

Finit is trained to skip sending SIGTERM prematurely to us, it skips all
processes with '@' in argv[0][0].  When Finit is half-way thru with the
shutdown it sends us SIGPWR and when it has completed it sends us the
SIGTERM to tell us to stop kicking the WDT to complete the reset.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>