If supervisor script returns OK (0), deregister process and continue
This patch adds support for the disarming, of sorts, of a triggered
supervisor event.  When a supervisor script is used, the script can
now return 0 to acknowledge and thereby disable the supervision of
an offending process.

E.g., a process suddenly misses a deadline.  The supervisor script can
be used to extend the functionality of watchdogd to, for example, try to
restart the process instead of reset the whole system.  When the script
returns zero watchdogd continues operating as per usual.

Or, instead of restarting the process, the script can write to a GPIO to
pull an external reset pin and then return non-zero to let watchdogd
save reset cause and reboot the CPU.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>