Fix #17: Change default behavior when supervisor is enabled
This patch changes the default behavior of watchdogd when the supervisor
is enabled.  As of now watchdogd unconditionally runs as SCHED_RR, with
the configured RT prio, not just when an external process is registered.

The change allows using a generic checker script to trigger an action
that is guaranteed to complete before a hardware WDT times out.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>