Clarify nomenclature; cause is from kernel, reason from watchdogd
Richard Alpe brought to my attention the confusing nomenclature used in
watchdogd today.  Reset *cause* and reset *reason* were sloppily, and
often interchangeably, used.  It was not clear to a user of watchdogd
what was the status from the kernel WDT, and what was the extra info
that watchdogd had added.

After a brief discussion he convinced me it was a good idea to clean
up this mess and we arreived at the following conclusion;

  - Reset cause should always be the `WDIOF_` flags returned from the
    kernel `WDIOC_GETBOOTSTATUS` ioctl.
  - Reset reason is the extra value watchdogd adds.  This, at its core,
    is a `code` that can be set by any of the supervisor or monitor
    plugins supported; e.g. a monitored process fails to meet its
    deadline, a system memory leak, or CPU overload (loadavg plugin)

95% of this patch is a search-and-replace of:

  - cause --> reason
  - cause --> code
  - etc.

The remaining 5% are adaptations, comments, and updated documentation.

The biggest implication of this change is the change in the libwdog API.
The wdog_reason_t::cause rename to ::code.  Hence the ABI version bump.
We could of course add an ugly union to support both ::cause and ::code,
but that seems silly since we might also want to add the kernel reset
cause to that struct later.  So, in the immortal words of a good friend
of mine;

>  "I'd rather be consistent than please everyone."

Also, I'd be very surpised if this really breaks anyone's code, because
that would mean someone other than Westermo actually uses libwdog ...

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>