Fix crash in case the "Label" or "Reset date" field is empty
chomp() on a empty string (strlen < 1) returns a NULL-ptr, which is not checked in strlcpy and leads to a crash.
The chomp here is not needed, because buf was chomped already before.
The string length check was added to make the following pointer operation safe (e.g. someone edits the file and deletes the space after the ':').
[Fix_watchdogd.state_parse_crash]