Address watchdogd service problem with generic script > 1s runtime.
This commit addresses a problem related to the watchdogd service. When trying to
activate a generic monitoring script with a runtime exceeding one second,
it triggers an unintended system reboot.

The error message reports:

"Monitor script PID <XXXX> is still running after <YY> seconds.
watchdog: watchdog0: watchdog did not stop!"

After thorough investigation, it was determined that the problem arises
from the fact that 'gs->script_runtime' is measured in milliseconds,
while 'gs->script_runtime_max' is maintained in seconds. This commit rectifies the issue.

Signed-off-by: senthil.thangaraj@getcruise.com