Refactor 0d73105, act on return of supervisor script immediately
This refactor adds optional callback support to script.c, where the
supervisor registers a callback to be run immediately upon exit of
the supervisor script.

There is no hard-coded 5 sec. timeout anymore, it is therefore up
to the administrator to ensure the script does not hang or block
the system in any way.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>