Workaround for libc differences
GLIBC and musl libc define __fsblkcnt64_t and __fsfilcnt64_t differntly.
The former define the, as long unsigned int, whereas the latter define
them as long long unsigned int.  Clearly we do not want to get involved
in this, so to work around the compiler warnings we cast to the larger
type and move on towards other challenges in life.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>