Update watchdogd status file when devices are changed at runtime
The status file can be read to determine last boot status and also the
currently running config, e.g., number of devices monitored etc.  So
when a user changes the the configuration this file should be updated
so the user can verify that the changes took.  This is also where any
adjusted timeout/interval from the config, based on feedback from the
kernel is stored.

Also, make sure fopen()/flose() are done in the same context.  Having
an implicit fclose() in wdt_fload_reason() was very sneaky and caused
a few hours of head scratching this afternoon.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>