Change safe-exit default value to true
Accidentally stopping (SIGTERM) watchdogd should not cause a reset of
the system by default.  It may be the safe choice, but for a regular
user, or distros upgrading watchdogd, it is not what is expected.

A professional (embedded) user will know to set this to false, so it
is a more suitable default to let this be true.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>